requirements.txt
Flask==2.2.2
numpy==1.22.4
Pillow==9.2.0
tensorflow==2.9.1
Werkzeug==2.2.2
gunicorn>=19.9.0